Stop depending on execution time for animations.

+++ core/src/me/scarlet/undertailor/gfx/Animation.java
@@ -102,15 +102,15 @@ public abstract class Animation<T extends KeyFrame> implements Disposable {
     public abstract Map<Long, T> getFrames();
     public abstract T getFrame(long stateTime, boolean looping);
     
-    public void drawFrame(long stateTime, boolean looping, String spriteset, float posX, float posY, float offX, float offY) {
+    public void drawFrame(double stateTime, boolean looping, String spriteset, float posX, float posY, float offX, float offY) {
         this.drawFrame(stateTime, looping, spriteset, posX, posY, offX, offY, 1F);
     }
     
-    public void drawFrame(long stateTime, boolean looping, String spriteset, float posX, float posY, float offX, float offY, float scale) {
+    public void drawFrame(double stateTime, boolean looping, String spriteset, float posX, float posY, float offX, float offY, float scale) {
         this.drawFrame(stateTime, looping, spriteset, posX, posY, offX, offY, scale, 0F);
     }
     
-    public abstract void drawFrame(long stateTime, boolean looping, String spriteset, float posX, float posY, float offX, float offY, float scale, float rotation);
+    public abstract void drawFrame(double stateTime, boolean looping, String spriteset, float posX, float posY, float offX, float offY, float scale, float rotation);
     
     public void finalChecks() {}
